Fix formatting of lint errors in GitHub Actions
This was - I think - broken in 7784c155f730a668ed38990002a8c47de58f163e,
when we changed our heredoc style. Previously, we had to escape shell
variables so that Docker _wouldn't_ substitute them, leaving them in
place for bash to substitute them. However, with new-style heredocs,
Docker doesn't do substitutions, so the escapes were left behind to be
consumed by bash.

Additionally, instead of using a formatter for eslint, we can use
"problem matchers" - i.e. regexes - which extract the relevant errors.
This works for tsc and stylelint in addition to eslint.