Allow top navbar to be scrolled away on most pages
The always-visible topbar is actually not the best use of space on most
pages:

* the topbar is easily found by scrolling back to the top of the page
* not having it take up 50 pixels of vertical space means 50 more rows
  for more-useful content like the puzzle list or guess queue
* having the top bar scroll away means we don't have to worry as much
  about having it be strongly-contrasting in color with other content
  that would slide under the fixed navbar.
* while some SPAs do have a more fixed layout, most other websites that
  have a navbar like this either do not make it sticky at all, or at
  least hide it when scrolling down

So let's just make the topbar be scrollable like most of the rest of the
web, and give a little more space for content.

We disable overscroll on pages where we use a fixed layout for the
majority of the page content and the navbar becomes the only part of the
page that uses default positioning, to avoid having just the navbar feel
wobbly.