Swap port allocations between mediasoup and coturn
MIT's network is documented to allow UDP connections on ports 50000+. By
moving mediasoup into that port range, we can hopefully avoid needing a
TURN relay on this network.

The deployment of these two changes will be a bit messy, as they can't
be pushed out atomically with each other. That should be fine as long as
they deploy when Jolly Roger isn't being heavily used.