Enable AWS IAM authentication with MongoDB.
- Use a fixed name for the role used to run EC2 instances so it can be
referenced in the MongoDB config. Note that if you try to set up JR
instances across multiple regions, it will now fail due to the role
name being duplicated. If this is a concern, we could append the region
name to the role (but this would require one user per region in
MongoDB).
- Include aws4 in dependencies as required for AWS authentication.