publishJoinedQuery enforces ordering with promises
This is all very subtle.

Before this patch, `publishJoinedQuery` set up new `observeChanges` live
queries in response to seeing an added() callback on another
`observeChanges` live query, and we relied upon the guarantee that all
initial `added()` callbacks on the new query would complete before the
`observeChanges()` returned its handle, to ensure that we would call
`this.sub.added()` on the targets of references before the objects
containing the id of the target, so that the client can safely assume
that any referenced ID is populated in the appropriate table.

Once we switch `observeChanges` to `observeChangesAsync`, however, we
lose that guarantee that we had previously relied upon.  We attempt to
recreate it explicitly with promise chains instead.

We want to ensure that when we incref(), if we created a new observer
(rather than incrementing the refcount of some existing one), we want to
avoid calling `.added()` on the sub (thus publishing the object to the
client) until that observer and any child observer(s) *it* creates,
recursively, has completed its initial `.added()` callbacks.  For
example, before publishing a Guess object to the client, we must publish
the User entry in the guess object's `.createdBy`, and the Puzzle object
in the guess object's `.puzzle`, and the Hunt object in the Puzzle
object's `.hunt` field.

So we make incref() itself async, returning a Promise<void> that will fulfill
only once any child observe queries are set up *and* have run through the
initial query results.

We should probably be careful about whether the refcount operations are
still safe in the presence of potential yield points.