Deduplicate new puzzles by URL.
When adding a puzzle, we grab a lock to atomically check if a puzzle
already exists with that URL and only add it if either no such puzzle
exists or if the client indicates that duplicate URLs are allowed. We
first do an opportunistic check to fail fast, and then grab the lock to
guard creating the Google Doc and tags before inserting the puzzle.
Empty URLs are never considered duplicates.

On the client side, we add a new checkbox to allow duplicate URLs,
unchecked and invisible by default. If we encounter a duplicate URL
error, we show an error indicating that the puzzle may be a dupe and
telling the user to check the now-visible checkbox if the duplicate URLs
are intentional.