Initial implementation of a browser extension to add puzzles.
Supports (at least) Chrome and Firefox.

Includes a settings page to input the Jolly Roger instance and API key,
and a popup menu which reads the URL and title from the current page and
provides a form to add it to a particular hunt on that instance. The
form is modeled off of PuzzleModalForm, but condensed vertically to try
to avoid scrolling in the height-limited popup window. We also surface
recently-used tags to make it easier to add multiple puzzles in the same
round/group.

Built as a mostly-independent module in the extension/ folder using npm;
could be split into a separate repository if desired. There is a small
bit of duplication with Jolly Roger proper, but this seems reasonable to
avoid unduly tight coupling and allow using more modern dependencies
than Meteor currently permits.

Uses React Bootstrap and consistent UI components with Jolly Roger, but
is powered by the REST API rather than Meteor, which seems like it would
introduce significant complexity for limited value. Also covered by the
root project's eslint and prettier configs; Typescript checking is
covered by the Webpack compilation. Some checks are disabled because
they require "npm install" to be run from the extension directory. In
general, we probably need to align on an organizational structure before
devoting too much time to making the linters work.