Update CF Lambdas from NodeJS v18 to v22.
v18 is slated to be deprecated on 9/1 and it will become impossible to
create new lambda functions with it on 10/1.

Tested by turning down and turning up stack and verifying lambdas
executed successfully. SshUsers wasn't triggering properly; as a
workaround, added a condition to only create the function/output when
serving is on (since the output is only needed by the EC2 template),
which ensures it is triggered upon turning serving off and on.