Deal with lying type definitions
These are places where the types as currently seen by tsc are incorrect,
because the type definitions do not match the actual code in question.
These can probably be unwound once we get to a world where we're using
first-party types for more things, but until then we have to deal with
the fact that the best-effort stuff from DefinitelyTyped no longer
matches our reality.

To unblock the broader Meteor upgrade, we assert a type, update some of
our addon type hints, and disable a couple lints.