Update lint:docs invocation to not error out
The `ts-node-esm` entrypoint does not seem to permit loading files with
the .mts file extension.  Instead, we can invoke node, using the
experimental --loader support, using the ts-node/esm loader, on the
check_docs.mts file to run the docs-freshness tests.

For what it's worth, I'm not sure the test itself does anything; I tried
making the claimed update date on docs/assets.md something ancient but
still got a 0 return code.  Some loglines indicated that it was at least
running the relevant checks against that file, so I'm just going to set
that aside as a peculiarity for later.