Update the Dockerfile for Meteor 3.2 and node 22
The new version of meteor objects to being run as superuser even if
you're just dispatching to other tools now, so we need more
`--allow-superuser`.

The new version of npm carried by Meteor 3.2 refuses to install from our
lockfile without `--legacy-peer-deps`, because react-split-pane declares
a maximum react version of 16 (and we're on react 18).

The new official instructions for the nodesource repos recommend
fetching and dearmoring the gpg key as described.