Replace react-split-pane/SplitPanePlus with SplitPaneMinus
`react-split-pane` is so old that it declares a react dependency that is
version-incompatible with what we actually use, and does not appear to
be likely to see an updated release.

In the interest of removing the --legacy-peer-deps requirement to
install anything (or update deps, or do other npm actions), this patch
reimplements the currently-used surface area of our react-split-pane
wrapper, `SplitPanePlus`, as `SplitPaneMinus`.

This does remove some latent unused functionality, like collapsible
panes and relative (percentage-based) resizing, but we can always
reimplement those features later if we decide we want to use them again
(and perhaps more easily than how they were done before, since we can
modify the underlying code instead of just wrapping it).

Unlike react-split-pane, this reimplementation is TypeScript-native.
Also unlike react-split-pane, we're willing to rely on PointerEvents
existing, so thanks to setPointerCapture, we can avoid writing all the
event handlers twice and still have them work on touch surfaces.