Reland: `meteor update` to 3.3.2 and enable SWC transpiler
This reverts commit 6594dcc279fd7d1f66823a60b5483608866719f6, which in
turn reverted 61d825eca7a37725c09123ad7ff3d2c74c4438a2.

In addition to the diff in 61d825eca7a37725c09123ad7ff3d2c74c4438a2, we
also remove the `--omit=optional` argument from our `npm install`
invocation in the built bundle directory in the production Docker image.
This is necessary because Meteor ships the whole babel-compiler package
as a dependency of the built production bundle, ostensibly to support
shell-server, but also because they never made that optional.

Remedying this does appear to be on upstream's radar, so perhaps we'll
have a better option around the 3.4 release.

[1] - https://github.com/meteor/meteor/issues/10902