Remove fetchAPIKey and rollAPIKey methods
The fetchAPIKey and rollAPIKey methods were originally designed in 2017
as a minimum-viable-product to be run from a browser JS console to allow
a puzzle-of-the-day system to use jolly-roger's email address registry
to restrict access.

These methods are now obviated by the createAPIKey and destroyAPIKey
methods and the apiKeysForSelf publication, which are more
fully-featured.