Record API key last-used time
When a key is successfully used to authenticate an API request, we now
record (or update) a `lastUsedAt` field to help make it clearer which
keys (if any) are in active or historical use.

To avoid API requests generating excessive writes, we eschew updating
`lastUsedAt` if we are still within 1 minute of the existing
`lastUsedAt` timestamp.  We can still potentially get a bit of a
thundering herd of attempted writes if there are many concurrent
API requests, but I don't think this is likely to be a real problem in
practice.