Switch to zodern:types for first-party Meteor types
This winds up pulling in a surprising amount of other stuff, especially
around things marked deprecated.  To wit: `fields` is now `projection`
in Mongo field specifiers, `WebApp.connectHandlers` is simply
`WebApp.handlers`, and findOne is deprecated on server, but
unfortunately the rule is a little loosely defined so we have to disable
it on client or else get a warning for every point load we do from
minimongo, which is annoying.

In addition to the obvious stuff, updated types around collections and
Mongo mean that our Zod-typed Model abstraction seems to be not working
quite as intended.  As written, some of the "is not a whole document
pattern" logic and type-fu meant to require that updateAsync() and
upsertAsync() receive at least one `$set` or similar modifier key
doesn't seem to work any more.  The runtime checks are still fine, but
it's a little less immediate feedback.

Also, a few things in `Accounts` are now defined as async functions
returning promises, but if you call them without a callback function in
client code, they block and the promise never arms, which is
non-functional, so I'm just voiding the results.

This does conveniently allow us to remove a few eslint-disables now that
the type definitions agree with the implementations.