Adapt to new mediasoup version
A few things moved around in the package, and a few places now need
slightly more precise typing to prove that field accesses are okay, and
now transports want you to name the protocol (TCP or UDP) you're
listening on.

I'm presuming that we probably prefer UDP connectivity for our
server-to-server pipe transports, but that we want to retain both UDP
and TCP options for client-to-server connections (because some firewalls
are restrictive, or we might have to fall back to a TURN server).