Replace react-copy-to-clipboard with a modern implementation
react-copy-to-clipboard has a couple issues that make me want to replace
it:

* it doesn't declare compatibility with React 19, so if we want to take
  the React 19 update we have to re-allow legacy peer deps
* the code to do so is merged upstream to master, but there's been no
  release in 3 years, and a related 9-month-old issue
* it doesn't have first-party typescript types
* also, it's ultimately just a thin wrapper around the copy-to-clipboard
  NPM package, which has its own cruft

When I look at the underlying copy-to-clipboard NPM package, I'm also
not super enthusiastic about it: it does a whole bunch of complicated
logic that affects the selection range, and then calls
`document.execCommand("copy")` (which is deprecated per MDN, which
recommends the use of the Clipboard API instead).  jolly-roger already
only supports the evergreen browsers, so there's no great reason for us
to use complicated logic for IE11 compatibility.

This PR implements a replacement copy-to-clipboard button component
which uses `navigator.clipboard.writeText()` to interact with the
clipboard, which is in the browser Baseline profile since 2020.

While I was here, I decided to also be consistent about including a
tooltip and feedback on successful copy, which fixes #2429.