Add helpful methods and publications
It seems reasonable to have a publication that allows retrieving the
Hunt associated with an InvitationCode, since you're going to have
access to it as soon as you join the hunt, and none of the fields of
Hunt are sensitive.  This also allows validating the invitation code
before prompting the user to join the hunt.

Slightly less obviously reasonably is a method that allows asking what
login methods a user account supports.  This is particularly useful to
avoid forcing the user to remember if they have an account provisioned
or not -- instead we can simply have them enter an email address and get
funneled into the appropriate flow.  To limit the information leakage
associated with this endpoint, it requires presenting an invitationCode,
and we return only "does this account exist" and "can it log in with a
password?" and "can it log in with a google account?"

It's worth noting that the password reset flow already leaks whether an
email address is associated with an account, by returning an error if
the email address is not associated with any account and succeeding
otherwise.  I think the value to users of being able to more reliably
access their accounts and avoiding sending them down the wrong flow
exceeds the risk of exposing this information.