Bump eslint dependency chain
Primarily a bump to the various typescript-eslint packages. As part of
that, typescript-eslint dropped support for declaring `tsconfigRootDir`
as "." in the eslint config, which means that we need an eslintrc.js
instead of a json file. Prettier ensures that it's formatted slightly
differently, but the content is mostly unchanged.

eslint-plugin-deprecation got rolled into typescript-eslint, so that's
dropped and replaced with the appropriate new overrides.

Most of the code changes are related to typescript-eslint being pickier
about enum type compatibility — it's now an error to compare an enum
value with a string or number directly, so those comparisons have been
tweaked.