Migrate eslint suppressions to biome
This is a good indicator of the type of stuff biome is less capable of
detecting than our current eslint config. Breaking down the rules that
we touch, there are some rules where Biome has an equivalent rule, there
are a few categories.

First, cases where there's an equivalent Biome rule, but it doesn't
catch the things we were suppressing for eslint, which seems fine:

* @typescript-eslint/no-use-before-define
* @typescript-eslint/consistent-type-imports => useImportType: This is
  triggering a bunch on our imports, and will need a cleanup run, but
  Biome seems fine with the places we were suppressing the eslint rule.
* @typescript-eslint/no-this-alias => noUselessThisAlias
* @typescript-eslint/no-unused-vars => noUnusedVariables
* react/no-array-index-key => noArrayIndexKey: This does still catch
  some issues, but seems more relaxed when the index is combined with
  other (presumably more unique) data.

And one case where Biome has an equivalent rule, but it doesn't seem to
trigger, and that seems suboptimal:

* @typescript-eslint/no-deprecated => noDeprecatedImports: This isn't
  triggering, likely because it's not resolving Meteor packages
  correctly.

There's also one case where I just opted not to enable the equivalent
rule because I thought it was unnecessarily restrictive:

* no-bitwise (enabled by eslint-config-airbnb)

For cases where there's no Biome equivalent, those can break down into
rules that (IMO) we can afford to lose:

* camelcase
* new-cap
* no-underscore-dangle
* filenames/match-exported
* import/no-mutable-exports
* import/first
* import/prefer-default-export: If anything, Biome seems to bias the
  other way with noDefaultExport, which I could get behind. (The
  argument is that editor auto-complete works better with named exports,
  which seems fair.)
* meteor/audit-argument-checks: Obviously this is Meteor-specific, so
  there's no Biome equivalent. Given that we're mostly migrated to typed
  publications and methods, I'm not overly concerned.
* react/destructuring-assignment: I like this rule, but we can live
  without it.

And finally, one rule that I want back:

* @typescript-eslint/prefer-nullish-coalescing: There's not currently a
  Biome equivalent, but we should keep an eye out for biomejs/biome#8043
  landing, since I'm a fan of this rule (I think nullish coalescing is
  generally much safer than its alternatives).