Replace Babel configuration with SWC config
After upgrading to Meteor 3, we switched to the "modern" build chain
which uses SWC instead of Babel, but didn't port over our configuration.

In the process, we lose the add-display-name plugin we were previously
using. It's unclear if it was still doing anything useful - most
components have display names set automatically. I tried using the
swc-plugin-add-display-name plugin but it doesn't seem to add display
names for any of the components currently missing them (mostly
components built from HOCs like React.memo).

As a final note, SWC does have a pretty narrow compatibility window with
its plugins, so we have to pull in a specific version that matches the
SWC version Meteor is using under the hood.