Upgrade discord.js to the latest version
This is the last package that had been blocked on updating due to Meteor
2's old Node version. Our interactions with the REST API work without
modifications, but we did need some changes to the websocket client.

Our previous strategy of waiting for the connection to initialize and
then probing the client's internal caches didn't seem to work this time
around, so instead handle events as they come in. The main issue is that
we need to make sure we clean up cache records for objects that were
deleted while we're offline, so keep track of that in-memory and clean
up once the connection is fully initialized.

Also, remove the old Discord discriminator field; the new client finally
removed it.