Enable noAccumulatingSpread
There's a good case to be made that in general this is an efficiency
concern (since each iteration requires generating a new accumulator and
copying the previous values). We were only failing this within JSON
schema generation, which isn't performance critical, but it's worth
porting that code over to have the generalized check.