Enable Biome useUniqueElementIds rule
While cleaning up the failures from this rule, I mostly took the
attitude that IDs don't need to be user-facing (with a few exceptions on
the setup page where it seemed useful to use them as anchors). The
necessary patches fall into a few categories.

First, a handful of static IDs seemed completely unused. I'm fairly
certain they were vestigial from before we converted to
styled-components, so I just removed them.

Second, the Tooltip component uses IDs to associate the tooltip with the
element being described (it sets aria-describedby behind the scenes). I
mostly eliminated where we were providing explicit IDs, instead just
using the `useId` hook.

Third, we use IDs to associate labels, help text, and form fields. I
switched all of these to use `useId` as a prefix, with a human-readable
suffix. Rather than pass an `id` attribute to `FormControl`,
react-bootstrap supports passing `controlId` to the enclosing
`FormGroup`, which it then propagates as appropriate to children. I
audited to ensure that we were consistently using that form (rather than
setting `htmlFor` on labels or an explicit `id` on the inputs).