Remove TODO for noImplicitAnyLet
The description of this rule sounds good, but doesn't seem to describe
the behavior of TypeScript in practice. It claims that uninitialized
variables declared with `let` or `var` but no type annotation will be
given the type `any`. But in practice, `tsc` seems to correctly infer
the type as `undefined` until a value is assigned.