Workers should clean themselves up if another server things they died
We've seen weird cases where a process's database records get garbage
collected while they are still running (likely due to scheduling
weirdness or temporary hangs). This causes problems as there can still
be in-memory resources (e.g. Mediasoup transports) that don't get
cleaned up and recreated when this happens. So if a server sees that it
has been garbage collected by another process, it should exit to try and
restore normalcy.

As a guard against aggressive cascading, only do this if the process is
part of a worker pool, and therefore will be replaced by the pool
manager.