Switch to tsgo for type checking
The go reimplementation of TypeScript is still a preview, but it claims
to be feature-complete with TypeScript 5.9 in all of the ways we care
about. (Most of the remaining deficiencies are around emitting
declarations and compiled code, and most of the behavioral changes are
around JSDoc comments; we use none of these features.)

And on the plus side, it's significantly faster, running on my laptop in
about 3s vs. 12s for tsc.

There were a few small changes required for compatibility:

* The "baseUrl" option in tsconfig.json is not supported. But the
  default value is fine. (I also cleaned up a no-longer-accurate
  comment)
* tsgo is pickier about default imports vs. namespace imports.
* tsgo is pickier about importing from files that aren't explicitly
  referenced in a package's "exports" declaration, so I reworded how we
  interact with audio level observers in mediasoup.

None of these changes regress running under tsc, so we can always switch
back.