Support @-mentioning operators in chat messages
This introduces the concept of a "role mention" as a separate construct
from a user "mention", although for the moment the only role that can be
mentioned is "operator".

To limit the blast radius of such mentions, @operator mentions only
notify operators that are currently online and with operating mode
enabled. This is detected by introducing a new usage of the subscribers
counters - operators with operating mode enabled for any hunt will
increment a "operators" subscriber counter with the hunts for which
they're actively operating in the context, allowing us to track them
down from the server.

If no operators are online with operating mode enabled, the mention goes
out to all operators for the hunt, active or not.