Rework the Discord websocket integration
The primary goal here was to make the flow of control clearer,
especially in cases where settings or feature flag changes trigger
restarts (or aborts) of the Discord gateway connection.

Additionally, the Discord client refresher was originally designed to
export a client object for other modules to use. But in practice, when
we want to talk to the Discord API outside of our gateway connection, we
do so through a separate Discord bot interface. So treat this module
more explicitly as a background daemon that runs as part of the overall
Jolly Roger process.

Rework Discord integration for flow