Drop most uses of Meteor.bindEnvironment
Prior to Meteor 3, using Meteor's mongo client required running code
within a fiber. Since non-Meteor libraries scheduling callbacks would do
so outside of a fiber, any time we passed a callback to such a library,
we had to wrap it in Meteor.bindEnvironment.

Under Meteor 3, this is no longer necessary, and Meteor.bindEnvironment
is only critical in scenarios where it's specifically important to
maintain Meteor's environment variables (which are largely connected to
whether code is currently running in the context of a method call or
publication for the purposes of tracking current connection and user).

Therefore, remove most uses of Meteor.bindEnvironment from the code.