Attempt to prevent zombie server processes
When a server has been garbage-collected by another server, we don't
want it to continue running, as we can no longer trust that the server's
in-process state is in sync with the database.

Therefore, when one server begins garbage-collecting another, it sets a
new `cleanupInProgressBy` field to indicate that the server should
self-terminate. Meanwhile, servers monitor their own record for that
field to get set (or the Server record to get deleted) and will exit if
it happens.