Fix two issues with haproxy config
First, add the alpn configuration to allow negotiating HTTP/2 traffic.
This would likely be the default behavior if we were running in HTTP
mode instead of TCP mode, but since we're in TCP mode we need to specify
it explicitly.

Second, don't try and connect to the HTTP/1 nginx backend in HTTP mode.
When nginx is re-populating its asset cache from the jolly-roger
backend, it sends the reply using HTTP/1.1 chunked encoding, and if
HAProxy tries to parse and relay that, something weird seems to happen
causing connections to fail. Since we don't actually need HAProxy to
*do* anything with the HTTP content, we can just leave it in TCP mode.

This seems to fix the intermittent hangs when first attempting to load
Jolly Roger.